refactor(podman): extract HyperVCheck class
