+++ extensions/podman/packages/extension/src/extension.ts
@@ -30,6 +30,7 @@ import { compareVersions } from 'compare-versions';
 import type { PodmanExtensionApi, PodmanRunOptions } from '../../api/src/podman-extension-api';
 import { SequenceCheck } from './checks/base-check';
 import { getDetectionChecks } from './checks/detection-checks';
+import { HyperVCheck } from './checks/hyperv-check';
 import { PodmanCleanupMacOS } from './cleanup/podman-cleanup-macos';
 import { PodmanCleanupWindows } from './cleanup/podman-cleanup-windows';
 import { KrunkitHelper } from './helpers/krunkit-helper';
@@ -42,7 +43,7 @@ import { getSocketCompatibility } from './utils/compatibility-mode';
 import type { InstalledPodman } from './utils/podman-cli';
 import { getPodmanCli, getPodmanInstallation } from './utils/podman-cli';
 import { PodmanConfiguration } from './utils/podman-configuration';
-import { HyperVCheck, PodmanInstall, WSL2Check, WSLVersionCheck } from './utils/podman-install';
+import { PodmanInstall, WSL2Check, WSLVersionCheck } from './utils/podman-install';
 import { ProviderConnectionShellAccessImpl } from './utils/podman-machine-stream';
 import { RegistrySetup } from './utils/registry-setup';
 import {
