+++ extensions/podman/packages/extension/src/utils/podman-install.spec.ts
@@ -26,14 +26,7 @@ import * as extensionObj from '../extension';
 import { releaseNotes } from '../podman5.json';
 import type { InstalledPodman } from './podman-cli';
 import type { Installer, PodmanInfo, UpdateCheck } from './podman-install';
-import {
-  getBundledPodmanVersion,
-  HyperVCheck,
-  PodmanInstall,
-  WinInstaller,
-  WSL2Check,
-  WSLVersionCheck,
-} from './podman-install';
+import { getBundledPodmanVersion, PodmanInstall, WinInstaller, WSL2Check, WSLVersionCheck } from './podman-install';
 import * as podmanInstallObj from './podman-install';
 import * as utils from './util';
 
@@ -586,196 +579,6 @@ test('expect winWSL2 command to be registered as disposable', async () => {
   // should contain a subscription with a disposable function
   expect(extensionContext.subscriptions[0]).toBeDefined();
   expect(extensionContext.subscriptions[0].dispose).toBeDefined();
-});
-
-describe('HyperV', () => {
-  beforeEach(() => {
-    vi.mocked(extensionApi.configuration.getConfiguration).mockReturnValue({
-      get: () => '',
-      has: vi.fn(),
-      update: vi.fn(),
-    });
-    vi.clearAllMocks();
-  });
-
-  test('expect HyperV preflight check return failure result if podman is older than 5.2.0', async () => {
-    let index = 0;
-    vi.spyOn(extensionApi.process, 'exec').mockImplementation(() => {
-      if (index++ < 1) {
-        return Promise.resolve({
-          stdout: 'podman version 5.1.0',
-          stderr: '',
-          command: 'command',
-        });
-      }
-      throw new Error();
-    });
-
-    const hyperVCheck = new HyperVCheck();
-    const result = await hyperVCheck.execute();
-    expect(result.successful).toBeFalsy();
-    expect(result.description).equal('Hyper-V is only supported with podman version >= 5.2.0.');
-  });
-
-  test('expect HyperV preflight skips podman version check if installationPreflightMode is true', async () => {
-    const execSpy = vi.spyOn(extensionApi.process, 'exec').mockImplementation(() => {
-      throw new Error();
-    });
-
-    const hyperVCheck = new HyperVCheck(true);
-    const result = await hyperVCheck.execute();
-    // exec should only be called once inside the isUserAdmin function, if so the podman check has been skipped as expected
-    expect(execSpy).toBeCalledTimes(1);
-    expect(execSpy).toBeCalledWith('powershell.exe', [
-      '$null -ne (whoami /groups /fo csv | ConvertFrom-Csv | Where-Object {$_.SID -eq "S-1-5-32-544"})',
-    ]);
-    expect(result.successful).toBeFalsy();
-  });
-
-  test('expect HyperV preflight check return failure result if it fails when checking admin user', async () => {
-    let index = 0;
-    vi.spyOn(extensionApi.process, 'exec').mockImplementation(() => {
-      if (index++ < 1) {
-        return Promise.resolve({
-          stdout: 'podman version 5.2.0',
-          stderr: '',
-          command: 'command',
-        });
-      } else {
-        throw new Error();
-      }
-    });
-
-    const hyperVCheck = new HyperVCheck();
-    const result = await hyperVCheck.execute();
-    expect(result.successful).toBeFalsy();
-    expect(result.description).equal('You must have administrative rights to run Hyper-V Podman machines');
-    expect(result.docLinks?.[0].url).equal(
-      'https://learn.microsoft.com/en-us/virtualization/hyper-v-on-windows/quick-start/enable-hyper-v',
-    );
-    expect(result.docLinks?.[0].title).equal('Hyper-V Manual Installation Steps');
-  });
-
-  test('expect HyperV preflight check return failure result if non admin user', async () => {
-    let index = 0;
-    vi.spyOn(extensionApi.process, 'exec').mockImplementation(() => {
-      if (index++ < 1) {
-        return Promise.resolve({
-          stdout: 'podman version 5.2.0',
-          stderr: '',
-          command: 'command',
-        });
-      } else {
-        return Promise.resolve({
-          stdout: 'False',
-          stderr: '',
-          command: 'command',
-        });
-      }
-    });
-
-    const hyperVCheck = new HyperVCheck();
-    const result = await hyperVCheck.execute();
-    expect(result.successful).toBeFalsy();
-    expect(result.description).equal('You must have administrative rights to run Hyper-V Podman machines');
-    expect(result.docLinks?.[0].url).equal(
-      'https://learn.microsoft.com/en-us/virtualization/hyper-v-on-windows/quick-start/enable-hyper-v',
-    );
-    expect(result.docLinks?.[0].title).equal('Hyper-V Manual Installation Steps');
-  });
-
-  test('expect HyperV preflight check return failure result if Podman Desktop is not run with elevated privileges', async () => {
-    let index = 0;
-    vi.spyOn(extensionApi.process, 'exec').mockImplementation(() => {
-      if (index++ <= 1) {
-        return Promise.resolve({
-          stdout: index === 1 ? 'podman version 5.2.0' : 'True',
-          stderr: '',
-          command: 'command',
-        });
-      } else {
-        throw new Error();
-      }
-    });
-
-    const hyperVCheck = new HyperVCheck();
-    const result = await hyperVCheck.execute();
-    expect(result.successful).toBeFalsy();
-    expect(result.description).equal(
-      'You must run Podman Desktop with administrative rights to run Hyper-V Podman machines.',
-    );
-    expect(result.docLinks).toBeUndefined();
-  });
-
-  test('expect HyperV preflight check return failure result if HyperV not installed', async () => {
-    let index = 0;
-    vi.spyOn(extensionApi.process, 'exec').mockImplementation(() => {
-      if (index++ <= 2) {
-        return Promise.resolve({
-          stdout: index === 1 ? 'podman version 5.2.0' : 'True',
-          stderr: '',
-          command: 'command',
-        });
-      } else {
-        throw new Error();
-      }
-    });
-
-    const hyperVCheck = new HyperVCheck();
-    const result = await hyperVCheck.execute();
-    expect(result.successful).toBeFalsy();
-    expect(result.description).equal('Hyper-V is not installed on your system.');
-    expect(result.docLinks?.[0].url).equal(
-      'https://learn.microsoft.com/en-us/virtualization/hyper-v-on-windows/quick-start/enable-hyper-v',
-    );
-    expect(result.docLinks?.[0].title).equal('Hyper-V Manual Installation Steps');
-  });
-
-  test('expect HyperV preflight check return failure result if HyperV not running', async () => {
-    let index = 0;
-    vi.spyOn(extensionApi.process, 'exec').mockImplementation(() => {
-      if (index++ <= 3) {
-        return Promise.resolve({
-          stdout: index === 1 ? 'podman version 5.2.0' : 'True',
-          stderr: '',
-          command: 'command',
-        });
-      } else {
-        throw new Error();
-      }
-    });
-
-    const hyperVCheck = new HyperVCheck();
-    const result = await hyperVCheck.execute();
-    expect(result.successful).toBeFalsy();
-    expect(result.description).equal('Hyper-V is not running on your system.');
-    expect(result.docLinks?.[0].url).equal(
-      'https://learn.microsoft.com/en-us/virtualization/hyper-v-on-windows/quick-start/enable-hyper-v',
-    );
-    expect(result.docLinks?.[0].title).equal('Hyper-V Manual Installation Steps');
-  });
-
-  test('expect HyperV preflight check return OK', async () => {
-    let index = 0;
-    vi.spyOn(extensionApi.process, 'exec').mockImplementation(() => {
-      if (index++ <= 4) {
-        return Promise.resolve({
-          stdout: index === 1 ? 'podman version 5.2.0' : index === 5 ? 'Running' : 'True',
-          stderr: '',
-          command: 'command',
-        });
-      } else {
-        throw new Error();
-      }
-    });
-
-    const hyperVCheck = new HyperVCheck();
-    const result = await hyperVCheck.execute();
-    expect(result.successful).toBeTruthy();
-    expect(result.description).toBeUndefined();
-    expect(result.docLinks?.[0].url).toBeUndefined();
-    expect(result.docLinks?.[0].title).toBeUndefined();
-  });
 });
 
 describe('getBundledPodmanVersion', () => {
